+++ test/test_sqs.py
@@ -433,6 +433,26 @@ class TestSQSHandlerRegistration:
             def handler_without_arguments():
                 return "one"
 
+    def test_bind_one_handler_to_multiple_events(self):
+        events = ["fooed", "bared"]
+
+        @SQSClient.handler(events)
+        def handler(*args, **kwargs):
+            print(args)
+            print(**kwargs)
+
+        for e in events:
+            assert e in SQSClient.handlers
+
+    def test_duplication_event_in_event_list(self):
+        events = ["spam", "spam"]
+
+        with pytest.raises(ValueError, match="Duplication found in list of event"):
+            @SQSClient.handler(events)
+            def handler(*args, **kwargs):
+                print(args)
+                print(**kwargs)
+
 
 class TestClientCreateMessage:
     @pytest.fixture(scope='function')